optimize: move missing enabled networks addresses alert from copy address trigger to all networks trigger
